Made some changes in Documents
